+++ TD1_3.vcxproj
@@ -147,9 +147,8 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Collision2D.cpp" />
-    <ClCompile Include="DebugText.cpp" />
     <ClCompile Include="include\engine\2D\Camera2D.cpp" />
+    <ClCompile Include="include\engine\2D\DebugText.cpp" />
     <ClCompile Include="include\engine\2D\Sprite.cpp" />
     <ClCompile Include="include\engine\3D\Camera3D.cpp" />
     <ClCompile Include="include\engine\3D\CameraTransform.cpp" />
@@ -165,23 +164,24 @@
     <ClCompile Include="include\engine\input\Input.cpp" />
     <ClCompile Include="include\engine\math\MathUtility.cpp" />
     <ClCompile Include="include\engine\math\Matrix.cpp" />
+    <ClCompile Include="include\engine\math\Quaternion.cpp" />
     <ClCompile Include="include\engine\math\Shape.cpp" />
     <ClCompile Include="include\engine\math\Vector.cpp" />
     <ClCompile Include="include\engine\PipelineFunc.cpp" />
+    <ClCompile Include="include\engine\TimeMamager.cpp" />
     <ClCompile Include="include\game\Block.cpp" />
+    <ClCompile Include="include\game\Collision2D.cpp" />
     <ClCompile Include="include\game\Game.cpp" />
     <ClCompile Include="include\game\GameScene.cpp" />
     <ClCompile Include="include\game\MainCamera.cpp" />
     <ClCompile Include="include\game\Resource.cpp" />
     <ClCompile Include="include\game\SceneCommonData.cpp" />
     <ClCompile Include="include\game\SceneManager.cpp" />
+    <ClCompile Include="include\game\SnowBall.cpp" />
     <ClCompile Include="include\game\Stage.cpp" />
     <ClCompile Include="include\game\TestObj.cpp" />
     <ClCompile Include="include\game\TitleScene.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Quaternion.cpp" />
-    <ClCompile Include="SnowBall.cpp" />
-    <ClCompile Include="TimeMamager.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="DirectXTex\DirectXTex_Desktop_2022_Win10.vcxproj">
@@ -219,9 +219,8 @@
     </FxCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Collision2D.h" />
-    <ClInclude Include="DebugText.h" />
     <ClInclude Include="include\engine\2D\Camera2D.h" />
+    <ClInclude Include="include\engine\2D\DebugText.h" />
     <ClInclude Include="include\engine\2D\Sprite.h" />
     <ClInclude Include="include\engine\3D\Camera3D.h" />
     <ClInclude Include="include\engine\3D\CameraTransform.h" />
@@ -250,8 +249,10 @@
     <ClInclude Include="include\engine\math\Vector3.h" />
     <ClInclude Include="include\engine\math\Vector4.h" />
     <ClInclude Include="include\engine\PipelineFunc.h" />
+    <ClInclude Include="include\engine\TimeManager.h" />
     <ClInclude Include="include\game\BaseScene.h" />
     <ClInclude Include="include\game\Block.h" />
+    <ClInclude Include="include\game\Collision2D.h" />
     <ClInclude Include="include\game\Game.h" />
     <ClInclude Include="include\game\GameObject.h" />
     <ClInclude Include="include\game\GameScene.h" />
@@ -260,11 +261,10 @@
     <ClInclude Include="include\game\Resource.h" />
     <ClInclude Include="include\game\SceneCommonData.h" />
     <ClInclude Include="include\game\SceneManager.h" />
+    <ClInclude Include="include\game\SnowBall.h" />
     <ClInclude Include="include\game\Stage.h" />
     <ClInclude Include="include\game\TestObj.h" />
     <ClInclude Include="include\game\TitleScene.h" />
-    <ClInclude Include="SnowBall.h" />
-    <ClInclude Include="TimeManager.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
