+++ include/engine/math/Quaternion.cpp
@@ -1,0 +1,3 @@
+#include "Quaternion.h"
+
+const Quaternion Quaternion::Identity = { 0.0f,0.0f,0.0f,1.0f };
